+++ app/src/main/java/com/depromeet/sloth/ui/home/HomeActivity.kt
@@ -3,7 +3,7 @@ package com.depromeet.sloth.ui.home
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.depromeet.sloth.R
-import com.depromeet.sloth.data.db.PreferenceManager
+import com.depromeet.sloth.data.PreferenceManager
 import com.depromeet.sloth.databinding.ActivityHomeBinding
 import com.depromeet.sloth.ui.home.lessonlist.LessonListFragment
 import com.depromeet.sloth.ui.home.mypage.MypageFragment
