+++ app/src/main/java/com/depromeet/sloth/ui/home/LessonViewModel.kt
@@ -1,7 +1,7 @@
 package com.depromeet.sloth.ui.home
 
 import androidx.lifecycle.viewModelScope
-import com.depromeet.sloth.data.db.PreferenceManager
+import com.depromeet.sloth.data.PreferenceManager
 import com.depromeet.sloth.data.network.home.*
 import com.depromeet.sloth.ui.base.BaseViewModel
 import kotlinx.coroutines.*
