+++ app/src/main/java/com/depromeet/sloth/ui/update/UpdateLessonActivity.kt
@@ -14,7 +14,7 @@ import androidx.annotation.RequiresApi
 import androidx.appcompat.content.res.AppCompatResources
 import androidx.appcompat.widget.AppCompatButton
 import com.depromeet.sloth.R
-import com.depromeet.sloth.data.db.PreferenceManager
+import com.depromeet.sloth.data.PreferenceManager
 import com.depromeet.sloth.data.network.register.RegisterLessonRequest
 import com.depromeet.sloth.data.network.update.UpdateLessonRequest
 import com.depromeet.sloth.data.network.update.UpdateLessonState
